requirements.txt
numpy==1.26.4
torch==2.0.1+cu118
torchvision==0.15.2+cu118
opencv-python==4.8.1.78
scipy==1.11.4
scikit-image==0.22.0
pillow==11.3.0
realesrgan==1.0
basicsr==1.4.2
tqdm
aiogram
